modified reward-value-tests

+++ src/main/java/RewardValue.java
@@ -5,7 +5,7 @@ public class RewardValue {
     public RewardValue(double cashValue)
     {
         this.cashValue =cashValue;
-        this.milesValue=cashValue/0.0035;
+        this.milesValue=(int) cashValue/0.0035;
     }
     public RewardValue(int milesValue)
     {
@@ -18,6 +18,6 @@ public class RewardValue {
     }
     public double getMilesValue()
     {
-        return (int)milesValue;
+        return milesValue;
     }
 }
